+++ lib/ui/widgets/adaptive_content.dart
